+++ lib/src/module.dart
@@ -4,46 +4,63 @@ import 'package:stark/src/injector.dart';
 typedef T FactoryFunc<T>(Injector i);
 typedef T FactoryFuncParams<T>(Injector i, Map<String,dynamic> params);
 
-abstract class Bind<T> {
+class Bind<T> {
 
   final FactoryFuncParams<T> factoryFuncParams;
   final FactoryFunc<T> factoryFunc;
   final String name;
+  final Type type;
+  final String scope;
   final bool isSingleton;
 
-  Bind({this.factoryFuncParams, this.factoryFunc,this.name,this.isSingleton, });
+  Bind._internal({
+    this.factoryFuncParams,
+    this.factoryFunc,
+    this.name,
+    this.type,
+    this.scope,
+    this.isSingleton
+  });
 }
 
-class Single<T> extends Bind<T>{
-
-  Single(
-      FactoryFunc<T> factory,
-      {
-        String named,
-      }
-  ) : super(factoryFunc: factory, isSingleton: true, name: named);
-
-  Single.withParams(
-      final FactoryFuncParams<T> factory,
-      {
-        String named,
-      }
-  ): super(factoryFuncParams: factory, isSingleton: true, name: named);
+Bind single<T>(FactoryFunc<T> factory, {String named, String scoped}){
+  return Bind._internal(
+      factoryFunc: factory,
+      isSingleton: true,
+      name: named,
+      type: T,
+      scope:scoped
+  );
 }
 
-class Factory<T> extends Bind<T> {
-
-  Factory(
-      FactoryFunc<T> factory,
-      {
-        String named,
-      }
-  ) : super(factoryFunc: factory, isSingleton: false, name: named);
-
-  Factory.withParams(
-      final FactoryFuncParams<T> factory,
-      {
-        String named,
-      }
-  ) : super(factoryFuncParams: factory, isSingleton: false, name: named);
+Bind singleWithParams<T>(FactoryFuncParams<T> factory, {String named, String scoped}){
+  return Bind._internal(
+      factoryFuncParams: factory,
+      isSingleton: true,
+      name: named,
+      type: T,
+      scope:scoped
+  );
 }
+
+Bind factory<T>(FactoryFunc<T> factory, {String named, String scoped}){
+  return Bind._internal(
+      factoryFunc: factory,
+      isSingleton: false,
+      name: named,
+      type: T,
+      scope:scoped
+  );
+}
+
+Bind factoryWithParams<T>(FactoryFuncParams<T> factory, {String named, String scoped}){
+  return Bind._internal(
+      factoryFuncParams: factory,
+      isSingleton: false,
+      name: named,
+      type: T,
+      scope:scoped
+  );
+}
+
+