+++ lib/stark.dart
@@ -1,5 +1,7 @@
 library stark;
 
 export 'src/stark.dart';
+export 'src/scope.dart';
+export 'src/disposable.dart';
 export 'src/module.dart';
 export 'src/stark_exception.dart';